Fixed naming error in mktemp and mktemp format in ed
mktemp() had an array of valid characters for the generated filename.
It wasn't using them, instead opting to use characters 0 to however
many characters were in the array.  The function has now been modified
to look up the valid character in the array and use that instead.

ed.c only had 5 X's in the mktemp format, and the minimum required
is 6.  The filename has also been cleaned up a little by changing
the prefix from 'e' to 'ed.'.